Added validation of BaseModel subclasses. Thanks to the same guy from earlier, I got some information on how `sanic` handles type enforcement. In FastAPI, this enforcement comes out-of-the-box, ready for your application, but in sanic, you have to explicitly state what aspect of the request you want type checked.
That wasn't too difficult. I had to then go through and strip all the responses out so that I can rewrite them correctly. Bit of a pain, but I'm incredibly grateful that it's all sorted now.

If you read through the code, you may see a few stray TODO's lying around. That's because I'm waiting to see if that section of code is really necessary, considering the fact that the `BaseModel` type enforcement _should_ have that as a feature. Again, I'm not sure - so we'll have to see